Calculate action consequence data for solution

This function calculates the consequences for each feature within a solution. Specifically, it shows how each action contributes to the overall amount of each feature given the solution.

Usage

solution_action_consequence(feature_ids, action_ids, pu_data, solution_data)

Arguments

feature_ids

character vector of identifiers for each feature.

action_ids

character vector of identifiers for each action.

pu_data

data.frame containing planning unit data.

solution_data

data.frame object containing the solution.

Value

A matrix containing the expected amount of each feature resulting from each action in the solution. Each row is a different feature, each column is a different action.