Update solution settings widget

Change the settings for solution settings widget on the client. Specifically, change the settings for a theme or weight depicted in a solution settings widget.

updateSolutionSettings(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  value
)

Arguments

session

The session object passed to function given to shinyServer Default is shiny::getDefaultReactiveDomain().

inputId

character The identifier of the input object.

value

list Object containing the new settings (see Details).

Details

Overview

The argument to value should be list object containing the information necessary to update a theme or weight. Broadly speaking, it should contain the following elements:

id

character value with the identifier for the theme or weight.

setting

character value with the name of the setting to update.

value

ANY new value for the setting.

type

character value indicating if the setting is a theme or weight.

Note that the value element in the list object should have a class (i.e. numeric, logical, or character) that is relevant to the setting that should be updated. For example, if the settingelement is equal toname, then the valueelement should contain acharacter` value. For reference, we provide examples detailing all the various settings that can be updated below.

Themes

Here we detail all possible settings that can be updated for themes.

Update the name. This controls the bold text shown in the header of the theme.

id

"THEMEID"

setting

"name"

value

"SPECIES"

type

"theme"

Update the status. This controls the large switch shown in the header of theme. Note that the value element must be a logical (TRUE or FALSE) value.

id

"THEMEID"

setting

"status"

value

TRUE

type

"theme"

Update the feature current values. This controls the current expected amount for a feature.

id

"THEMEID"

setting

"feature_current"

value

0.1

type

"theme"

Update the feature goals. This controls the slider used to specify goal for a feature.

id

"THEMEID"

setting

"feature_goal"

value

0.8

type

"theme"

Update the feature maxima. This controls the maximum possible amount for each feature.

id

"THEMEID"

setting

"feature_max"

value

34

type

"theme"

Weights

Here we detail all possible settings that can be updated for weights.

Update the name. This controls the bold text shown in the header.

id

"WEIGHTID"

setting

"name"

value

"SPECIES"

type

"weight"

Update the factor. This controls the slider shown for the feature.

id

"WEIGHTID"

setting

"factor"

value

0.1

type

"weight"

Parameter

Here we detail all possible settings that can be updated for the parameter.

Update the parameter. This controls the highest possible value for the slider.

id

"PARAMETERID"

setting

"value"

value

9999

type

"parameter"

See also