whattodo: Interactive management action prioritization application

The What To Do application is a decision support tool to help prioritize management actions for the Nature Conservancy of Canada. Data can be uploaded using an Excel Spreadsheet and (optional) a shapefile delineating the spatial location of sites. Prioritizations are generated using mixed integer programming techniques. The performance of candidate prioritizations can be examined using summary statistics and tables. Finally, data and prioritizations can also be downloaded for subsequent analysis.

See also

Author

Maintainer: Jeffrey O Hanson jeffrey.hanson@uqconnect.edu.au (ORCID)

Authors:

Richard Schuster richard.schuster@glel.carleton.ca (ORCID)

Joseph Bennett

Jaimie Vincent

Other contributors:

Dan Wismer dan.wismer@natureconservancy.ca [contributor]

Examples

# \donttest{
if (interactive()) {
run_app()
}
# }