New project

Create a new Project object.

Usage

new_project(
  site_ids,
  site_descriptions,
  feature_ids,
  feature_descriptions,
  action_ids,
  action_descriptions,
  site_data,
  feature_data,
  feasibility_data,
  consequence_data,
  parameters,
  site_geometry = NULL,
  id = uuid::UUIDgenerate()
)

Arguments

site_ids

character identifiers for sites. No missing (NA) values are permitted. This object must contain at least one value.

site_descriptions

character descriptions of sites. No missing (NA) values are permitted. This object must contain at least one value.

feature_ids

character identifiers for biodiversity features. No missing (NA) values are permitted. This object must contain at least one value.

feature_descriptions

character descriptions of biodiversity features. No missing (NA) values are permitted. This object must contain at least one value.

action_ids

character identifiers for management actions. No missing (NA) values are permitted. This object must contain at least one value.

action_descriptions

character descriptions of management actions. No missing (NA) values are permitted. This object must contain at least one value.

site_data

data.frame containing site data.

feature_data

data.frame containing feature data.

feasibility_data

data.framecontaining feasibility data.

consequence_data

list of data.frame objects containing consequence data.

parameters

list object containing parameters to customize appearance of worksheet.

site_geometry

sf::st_sf() object containing spatial boundaries for the sites. Defaults to NULL such that site locations are generated based on the longitude/latitude of each site (per argument to site_data).

id

character identifier for object.

Value

A Project object.

Details

This function creates an Excel Workbook with data and results.

Examples

#TODO